# rest/docs/swagger.yaml
basePath: /
definitions:
...
  api.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  api.PagedResults:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.AuthToken:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      expiry:
        type: string
      id:
        type: integer
      token:
        type: string
      updated_at:
        type: string
      upload_only:
        type: boolean
      user:
        type: integer
    type: object
  model.Autoretrieve:
    properties:
      addresses:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      handle:
        type: string
      id:
        type: integer
      last_connection:
        type: string
      peer_id:
        type: string
      token:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3030
info:
  contact:
    email: me@me.com
    name: Me
    url: http://me.com/terms.html
  description: Sample CRUD objects-api for estuary db
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Sample CRUD objects-api for estuary db
  version: "1.0"
paths:
  /authtokens:
    get:
      consumes:
      - application/json
      description: GetAllAuthTokens is a handler to get a slice of record(s) from
        auth_tokens table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AuthToken'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of AuthTokens
      tags:
      - AuthTokens
    post:
      consumes:
      - application/json
      description: add to add a single record to auth_tokens table in the estuary
        database
      parameters:
      - description: Add AuthTokens
        in: body
        name: AuthTokens
        required: true
        schema:
          $ref: '#/definitions/model.AuthToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AuthToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to auth_tokens table
      tags:
      - AuthTokens
  /authtokens/{argID}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from auth_tokens table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.AuthToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from auth_tokens
      tags:
      - AuthTokens
    get:
      consumes:
      - application/json
      description: GetAuthTokens is a function to get a single record from the auth_tokens
        table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AuthToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table AuthTokens by  argID
      tags:
      - AuthTokens
    put:
      consumes:
      - application/json
      description: Update a single record from auth_tokens table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      - description: Update AuthTokens record
        in: body
        name: AuthTokens
        required: true
        schema:
          $ref: '#/definitions/model.AuthToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AuthToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table auth_tokens
      tags:
      - AuthTokens
  /autoretrieves:
    get:
      consumes:
      - application/json
      description: GetAllAutoretrieves is a handler to get a slice of record(s) from
        autoretrieves table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Autoretrieve'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of Autoretrieves
      tags:
      - Autoretrieves
    post:
      consumes:
      - application/json
      description: add to add a single record to autoretrieves table in the estuary
        database
      parameters:
      - description: Add Autoretrieves
        in: body
        name: Autoretrieves
        required: true
        schema:
          $ref: '#/definitions/model.Autoretrieve'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Autoretrieve'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to autoretrieves table
      tags:
      - Autoretrieves
  /autoretrieves/{argID}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from autoretrieves table in the estuary
        database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Autoretrieve'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from autoretrieves
      tags:
      - Autoretrieves
    get:
      consumes:
      - application/json
      description: GetAutoretrieves is a function to get a single record from the
        autoretrieves table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Autoretrieve'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table Autoretrieves by  argID
      tags:
      - Autoretrieves
    put:
      consumes:
      - application/json
      description: Update a single record from autoretrieves table in the estuary
        database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      - description: Update Autoretrieves record
        in: body
        name: Autoretrieves
        required: true
        schema:
          $ref: '#/definitions/model.Autoretrieve'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Autoretrieve'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table autoretrieves
      tags:
      - Autoretrieves
...
swagger: "2.0"